refactor: split workflow studio page
